revert(board): drop the triggeringMessage bolt-on for the events rail
Reverts df8bf96 + cd243b3. Carrying the triggering message on conversation:*
and appending it to a recentMessages preview cache was a parallel message path
mimicking db.events. Now that the board card reads bodies from the events rail
(prior commit), that bolt-on is dead: the backend payload field, both emit-site
populations, the board-post-message projection extraction, and the frontend
merge-append all go. The cached projection on the conversation row remains as the
card's cold-start fallback, seeded by the board fetch like before.
