+++ apps/frontend/src/stores/board-store.test.ts
@@ -6,16 +6,12 @@ import type { BoardPost, BoardPostMessage, ConversationWithStaleness } from "@th
 
 const WORKSPACE_ID = "ws_1"
 
-function makeConversation(
-  id: string,
-  lastActivityAt: string,
-  messageIds: string[] = ["m1"]
-): ConversationWithStaleness {
+function makeConversation(id: string, lastActivityAt: string): ConversationWithStaleness {
   return {
     id,
     streamId: "stream_1",
     workspaceId: WORKSPACE_ID,
-    messageIds,
+    messageIds: ["m1"],
     participantIds: ["usr_1"],
     secondaryMessageIds: [],
     topicSummary: id,
@@ -31,7 +27,7 @@ function makeConversation(
   }
 }
 
-function makeMessage(id: string, createdAt = "2026-06-22T12:00:00.000Z"): BoardPostMessage {
+function makeMessage(id: string): BoardPostMessage {
   return {
     id,
     authorId: "usr_1",
@@ -40,13 +36,13 @@ function makeMessage(id: string, createdAt = "2026-06-22T12:00:00.000Z"): BoardP
     reactions: {},
     attachments: [],
     linkPreviews: [],
-    createdAt,
+    createdAt: "2026-06-22T12:00:00.000Z",
   }
 }
 
 function makePost(id: string, lastActivityAt: string, recentMessages: BoardPostMessage[] = []): BoardPost {
   return {
-    conversation: makeConversation(id, lastActivityAt, ["m1", ...recentMessages.map((m) => m.id)]),
+    conversation: makeConversation(id, lastActivityAt),
     openingMessage: makeMessage("m1"),
     recentMessages,
     totalReplies: recentMessages.length,
@@ -88,92 +84,12 @@ describe("mergeBoardConversation", () => {
       makePost("conv_1", "2026-06-20T12:00:00.000Z", [makeMessage("r1")]),
       makePost("conv_2", "2026-06-21T12:00:00.000Z"),
     ])
-    const merged = await mergeBoardConversation(
-      "conv_1",
-      makeConversation("conv_1", "2026-06-22T12:00:00.000Z", ["m1", "r1"])
-    )
+    const merged = await mergeBoardConversation("conv_1", makeConversation("conv_1", "2026-06-22T12:00:00.000Z"))
     expect(merged).toBe(true)
     const board = await readBoard()
     expect(board.map((p) => p.id)).toEqual(["conv_1", "conv_2"])
-    // No triggering body on this event — the existing preview is preserved.
+    // The event carries the aggregate, not message bodies — preview is preserved.
     expect(board[0]?.recentMessages.map((m) => m.id)).toEqual(["r1"])
-  })
-
-  it("appends the triggering message body so a new reply shows in place", async () => {
-    await seedBoardPosts(WORKSPACE_ID, [makePost("conv_1", "2026-06-20T12:00:00.000Z", [makeMessage("r1")])])
-    await mergeBoardConversation(
-      "conv_1",
-      makeConversation("conv_1", "2026-06-22T12:00:00.000Z", ["m1", "r1", "r2"]),
-      makeMessage("r2", "2026-06-22T12:00:00.000Z")
-    )
-    const row = await db.conversations.get("conv_1")
-    expect(row?.recentMessages.map((m) => m.id)).toEqual(["r1", "r2"])
-    expect(row?.totalReplies).toBe(2)
-  })
-
-  it("dedups the triggering body against an optimistic write or a re-echo", async () => {
-    await seedBoardPosts(WORKSPACE_ID, [makePost("conv_1", "2026-06-20T12:00:00.000Z", [makeMessage("r1")])])
-    const echo = () =>
-      mergeBoardConversation(
-        "conv_1",
-        makeConversation("conv_1", "2026-06-22T12:00:00.000Z", ["m1", "r1"]),
-        makeMessage("r1", "2026-06-22T12:00:00.000Z")
-      )
-    await echo()
-    await echo()
-    const row = await db.conversations.get("conv_1")
-    expect(row?.recentMessages.map((m) => m.id)).toEqual(["r1"])
-    expect(row?.totalReplies).toBe(1)
-  })
-
-  it("orders the appended body by time and caps the preview at three", async () => {
-    await seedBoardPosts(WORKSPACE_ID, [
-      makePost("conv_1", "2026-06-20T12:00:00.000Z", [
-        makeMessage("r1", "2026-06-22T12:00:01.000Z"),
-        makeMessage("r2", "2026-06-22T12:00:02.000Z"),
-        makeMessage("r3", "2026-06-22T12:00:03.000Z"),
-      ]),
-    ])
-    await mergeBoardConversation(
-      "conv_1",
-      makeConversation("conv_1", "2026-06-22T12:00:04.000Z", ["m1", "r1", "r2", "r3", "r4"]),
-      makeMessage("r4", "2026-06-22T12:00:04.000Z")
-    )
-    const row = await db.conversations.get("conv_1")
-    expect(row?.recentMessages.map((m) => m.id)).toEqual(["r2", "r3", "r4"])
-    expect(row?.totalReplies).toBe(4)
-  })
-
-  it("recomputes the reply count from the aggregate when a message is reassigned away", async () => {
-    await seedBoardPosts(WORKSPACE_ID, [
-      makePost("conv_1", "2026-06-20T12:00:00.000Z", [makeMessage("r1"), makeMessage("r2")]),
-    ])
-    // No triggering body: a reassignment shrank the conversation's messageIds.
-    await mergeBoardConversation("conv_1", makeConversation("conv_1", "2026-06-22T12:00:00.000Z", ["m1", "r1"]))
-    const row = await db.conversations.get("conv_1")
-    expect(row?.totalReplies).toBe(1)
-  })
-
-  it("preserves the seeded reply count when the opening message was deleted (ambiguous slice)", async () => {
-    // openingMessage null → the flat-vs-thread slice can't be determined, so the
-    // last server-computed totalReplies must survive rather than flip off-by-one.
-    await seedBoardPosts(WORKSPACE_ID, [
-      { ...makePost("conv_1", "2026-06-20T12:00:00.000Z", [makeMessage("r1")]), openingMessage: null },
-    ])
-    await mergeBoardConversation("conv_1", makeConversation("conv_1", "2026-06-22T12:00:00.000Z", ["m1", "r1"]))
-    const row = await db.conversations.get("conv_1")
-    expect(row?.totalReplies).toBe(1)
-  })
-
-  it("ignores a triggering body that isn't a primary reply of this conversation", async () => {
-    await seedBoardPosts(WORKSPACE_ID, [makePost("conv_1", "2026-06-20T12:00:00.000Z", [makeMessage("r1")])])
-    await mergeBoardConversation(
-      "conv_1",
-      makeConversation("conv_1", "2026-06-22T12:00:00.000Z", ["m1", "r1"]),
-      makeMessage("secondary_x", "2026-06-22T12:00:00.000Z")
-    )
-    const row = await db.conversations.get("conv_1")
-    expect(row?.recentMessages.map((m) => m.id)).toEqual(["r1"])
   })
 
   it("returns false when the card isn't cached (caller hydrates instead)", async () => {
@@ -189,11 +105,7 @@ describe("mergeBoardConversation", () => {
     await seedBoardPosts(WORKSPACE_ID, [makePost("conv_1", "2026-06-20T12:00:00.000Z")])
     await optimisticBoardReply("conv_1", makeMessage("r1"), Date.parse("2026-06-21T12:00:00.000Z"))
     expect((await db.conversations.get("conv_1"))?._status).toBe("pending")
-    await mergeBoardConversation(
-      "conv_1",
-      makeConversation("conv_1", "2026-06-21T12:00:05.000Z", ["m1", "r1"]),
-      makeMessage("r1")
-    )
+    await mergeBoardConversation("conv_1", makeConversation("conv_1", "2026-06-21T12:00:05.000Z"))
     expect((await db.conversations.get("conv_1"))?._status).toBeUndefined()
   })
 })
